Semi working AI
AI seems to actually change button labels but only when the button is clicked. TODO: make code that changes the value of buttons when the player clicks the button once, and changes the buttons according to the RNG output without the player having to click it. Create a line checker that checks line combinations and adds a point to score board. Create button code that clears scoreboard.