minor changes, script still needs evaluation.
computer player still chooses players squares and player is still able to click their own squares twice, game still doesnt plot an X but only plots an O when the player clicks a square at the start, and code for rng needs to be implemented to prevent the rng from choosing the same square twice but this can be done with an array but will need a large portion of the code to be reworked.

TO DO: fix all issues mentioned above and any other issues that may appear, finish the charts, documentation etc.